refactor : 소셜 로그인 패키지 수정
- jpa 오류로 root build.gradle에 추가
- domain -> api로 패키지 이동
