+++ meerket/build.gradle
@@ -35,6 +35,7 @@ subprojects {
     }
 
     dependencies {
+        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
         implementation 'com.fasterxml.jackson.core:jackson-databind:2.18.1'
         implementation 'org.springframework.boot:spring-boot-starter-validation'
         compileOnly 'org.projectlombok:lombok'
